Finish Batching

+++ process_mnist.py
@@ -30,23 +30,24 @@ with gzip.open(data_path + 'mnist_test.csv.gz', 'rt') as file:
         test_images.append(line[1:])
 
 # Convert to numpy arrays
-train_images = np.array(train_images, dtype='int')
-train_labels = np.array(train_labels, dtype='int')
-test_images = np.array(test_images, dtype='int')
-test_labels = np.array(test_labels, dtype='int')
+train_images = np.array(train_images, dtype='float32')
+train_labels = np.array(train_labels, dtype='float32')
+test_images = np.array(test_images, dtype='float32')
+test_labels = np.array(test_labels, dtype='float32')
 
 # Calculate number of classes
-# num_classes = len(np.unique(train_y))
+# num_classes = len(np.unique(train_labels))
 #
 # # Convert labels to one-hot encodings
-# train_y = np.eye(num_classes, dtype='int')[train_y]
-# test_y = np.eye(num_classes, dtype='int')[test_y]
+# train_labels = np.eye(num_classes, dtype='float32')[train_labels]
+# test_labels = np.eye(num_classes, dtype='float32')[test_labels]
 
 # Pickle
-data = {'train_images': train_images, 'train_labels': train_labels, 'test_images': test_images, 'test_labels': test_labels}
+data = {'train_images': train_images,
+        'train_labels': train_labels,
+        'test_images': test_images,
+        'test_labels': test_labels}
+
 with gzip.open(data_path + 'mnist.pkl', 'wb') as file:
     pickle.dump(data, file)
-    # pickle.dump(train_images, file)
-    # pickle.dump(train_labels, file)
-    # pickle.dump(test_images, file)
-    # pickle.dump(test_labels, file)
+
